AI-Driven Nutritional Analysis and SQL Query Generator/sqlllm/requirements.txt
streamlit==1.10.0
google-generativeai==0.0.2
python-dotenv==0.19.2
langchain==0.5.0
PyPDF2==1.26.0
chromadb==0.1.1
faiss-cpu==1.7.1
pdf2image==1.16.0
sqlite3==2.6.0